Work session 1
Wed Dec 16 10:30 am - 12pm
The first four problems are fairly simple, so I had a lot of successes. I'm confident that my code produces the correct answers for these four problems. In problems 1 and 2, I mainly used for and while loops and did not include any function as these problems were very simple. I only started to implement some functions in problems 3 and 4. In problem 3, I created a find_factor function to find the largest factor of a certain number. There are if statements in this function that check whether a number is a factor. In problem 4, I created check_pallindrome function to check if a number is a pallindrome. This function checks if the first half of a number is similar to the second half of that number.