Work session 2
Jan 9th, 10 pm - 11:20 pm 

During this work session, I have completed problems 5,6,7 and 8. In problem 5, I have found the correct answer but my code runs very slowly as it was mostly a brute force algorithm; my code essentially just adds 1 to the current number until it finds the right number. So I'm planning to come back to this problem. Problem 6 was very straight forward as I only had to use two different for loops to find the sum of the squares of the first one hundred natural numbers and the square of the sum of the first one hundred natural numbers. For problem 7, I decided to implement two functions: a function to check whether a number is a prime number and a function to find nth prime number. In problem 8, I think my code produces the correct answer. However, my approach is not very efficient as I had to create thirteen different variables inside my while loop for thirteen different numbers.