Work session 3
Wed Jan 13, 2020 1:30 - 3 pm
 
During this work session, I complete problems 9, 10, and some of 11. For problem 9, I used two for loops and two if statements to find the value of a, b, and c. I'm certain that I have found the correct answer for this problem. However, since this my solution for this problem was a brute force algorithm, the program takes a while to finish. For problem 10, I was able to reuse my primecheck function from problem 7 to check whether a number is a prime number.This function made it very easy to find all the primes below 2 million. However, my code for this problem was also a brute force algorithm so it took very long to load. For problem 11, I have only been able to create a matrix of the numbers and a function to find to product of numbers horizontally.