Work session 4
Sun Jan 17 2:20 am - 3:20 am 
During this work session, I have completed problems 11, 12 and 13. For problem 11, I finished writing functions for calculating products of numbers diagonally and vertically. With these functions and for loops, I was able to calculate different possible largest products for the matrix. However, after checking my answer, It took me a while to realize that my code did not include some numbers in the matrix and debug my program. For problem 12, I created a function that counts the number of divisions for a certain number and a function that generates a triangle number.  I used a while loop to keep adding one to count until the program finds a triangle number that has over 500 divisors. Since this is a brute force approach, my code takes very long to execute. Problem 13 was fairly straight forward. I first put all of the numbers in the string into a list and found their sum by using a for loop.